+++ src/Altcha.Net/AltchaChallenge.cs
@@ -1,5 +1,6 @@
-using System.Text.Json;
+#if !NET48
 using System.Text.Json.Serialization;
+#endif
 
 namespace Altcha.Net;
 
@@ -14,23 +15,33 @@ public sealed class AltchaChallenge
         MaxNumber = maxNumber;
     }
 
+#if !NET48
     [JsonPropertyName("algorithm")]
+#endif
     public string Algorithm { get; }
 
+#if !NET48
     [JsonPropertyName("challenge")]
+#endif
     public string Challenge { get; }
 
+#if !NET48
     [JsonPropertyName("salt")]
+#endif
     public string Salt { get; }
 
+#if !NET48
     [JsonPropertyName("signature")]
+#endif
     public string Signature { get; }
 
+#if !NET48
     [JsonPropertyName("maxnumber")]
+#endif
     public int MaxNumber { get; }
 
     public string ToJson()
     {
-        return JsonSerializer.Serialize(this, AltchaJson.Options);
+        return AltchaJson.SerializeChallenge(this);
     }
 }